+++ app/helpers/subject_schedules_helper.rb
@@ -1,0 +1,2 @@
+module SubjectSchedulesHelper
+end