+++ app/models/schedule.rb
@@ -12,7 +12,29 @@ class Schedule < ActiveRecord::Base
   # Model Relationships
   belongs_to :user
   belongs_to :project
+  has_many :subject_schedules
 
   # Model Methods
+
+  def designs(ids)
+    self.project.designs.where( id: ids ).order( :name )
+  end
+
+  def sorted_items
+    self.items.sort{ |a,b| relative_size( a ) <=> relative_size( b ) }
+  end
+
+
+  private
+
+    def relative_size(item_hash)
+      interval = item_hash.symbolize_keys[:interval].to_i
+      units = item_hash.symbolize_keys[:units]
+      case units when 'business days'
+        interval.send('days')
+      else
+        interval.send(units)
+      end
+    end
 
 end