+++ db/migrate/20131003190259_create_subject_schedules.rb
@@ -1,0 +1,11 @@
+class CreateSubjectSchedules < ActiveRecord::Migration
+  def change
+    create_table :subject_schedules do |t|
+      t.integer :subject_id
+      t.integer :schedule_id
+      t.date :initial_due_date
+
+      t.timestamps
+    end
+  end
+end