+++ db/migrate/20131004135226_create_events.rb
@@ -1,0 +1,13 @@
+class CreateEvents < ActiveRecord::Migration
+  def change
+    create_table :events do |t|
+      t.string :name
+      t.text :description
+      t.integer :project_id
+      t.integer :user_id
+      t.boolean :deleted, null: false, default: false
+
+      t.timestamps
+    end
+  end
+end