+++ test/helpers/subject_schedules_helper_test.rb
@@ -1,0 +1,4 @@
+require 'test_helper'
+
+class SubjectSchedulesHelperTest < ActionView::TestCase
+end