+++ test/models/subject_schedule_test.rb
@@ -1,0 +1,7 @@
+require 'test_helper'
+
+class SubjectScheduleTest < ActiveSupport::TestCase
+  # test "the truth" do
+  #   assert true
+  # end
+end